rename constants-->types
